Convert the data file to utf-8 to avoid expensive entity replacement.
Previously the file was encoded as iso8859-1 and the non-iso characters
were encoded as entities like <ij>, that had to be replaced by their
equivalent in unicode when the detector was spawned. This operation was
very expensive and modern encodings like utf-8 allow saving the special
characters directly in the file, avoiding the extra initialization time.